requirements.txt
dbt-duckdb
duckdb
pandas
numpy
pyarrow
scikit-learn
xgboost
pyyaml
